Unlimited export
The export did not set a limit, so it was only 20 items long per default, which just does not make sense.
Maybe limiting the export would be a feature, but as long at this is not abailable, there should not be any limit.
Setting the limit to "0" gives you the whole thing. So you get what you ordered: an export of everything.