Fix Argument 2 must be of type string
If the option emailSelectToName is not set, the option emailSelectTo is used. emailSelectTo was exploded before it was assigned to emailSelectToName. After that it is exploded again which causes the following issue in PHP 8. The patch just reorders the assignment.

```
PHP Fatal error:  Uncaught TypeError: explode(): Argument #2 ($string) must be of type string, array given in /…/core/components/formit/src/FormIt/Hook/Email.php:99
```